chore: :construction: Temporarily replace ProductGalleryCloudinary
Swap out the ProductGalleryCloudinary component with ProductSlideshow to diagnose potential freezing issues in the app. This change aims to determine if the gallery component is causing performance problems or if the issue lies elsewhere in the application.

This change is part of troubleshooting efforts to identify the root cause of app freezing.